Coonsertando tipo integer para decumal

+++ ExConversorDeMedidas/sql/create_historic_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE historic(
     id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL UNIQUE,
-    value_from INTEGER NOT NULL,
+    value_from DECIMAL NOT NULL,
     unity_from VARCHAR NOT NULL,
-    value_to INTEGER NOT NULL,
+    value_to DECIMAL NOT NULL,
     unity_to VARCHAR NOT NULL
 )
